requirements.txt
azure-core
azure-ai-inference
python-dotenv
flask